+++ js/observer.js
@@ -1,8 +1,8 @@
 export const observer = new IntersectionObserver((entries) => {
     entries.forEach((entry) => {
         if (entry.isIntersecting) {
-            entry.target.classList.add('show');
-        }   else {
+            entry.target.classList.add('show')
+        } else {
             entry.target.classList.remove('show')
         }
     })